requirements.txt
# Core Web Framework
streamlit==1.28.1

# LangChain Framework
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0

# HuggingFace Integration (Active)
transformers==4.36.0
torch==2.1.0
langchain-huggingface==0.0.1
huggingface-hub==0.19.4
sentence-transformers==2.2.2
accelerate==0.24.1
tokenizers==0.15.0
safetensors==0.4.1

# OpenAI Integration (Commented out)
# langchain-openai==0.0.2
# openai==1.3.8

# PDF Processing
PyMuPDF==1.23.8
pdfminer.six==20221105

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Additional Dependencies
typing-extensions==4.8.0
pillow==10.1.0
markdown==3.5.1
regex==2023.10.3

# Optional: For better performance
# psutil==5.9.6  # System monitoring
# tqdm==4.66.1   # Progress bars